requirements.txt
# Python packages (system packages preferred — see install.sh)
# Pillow is only needed to generate icon.png; installed via apt as python3-pil
Pillow
